Continued work on move generation.
Pawn and knight moves finished. Added precomupted.py file to partially workout the knight moves from all squares on the board so that they can be acessed and less calculation is needed during runtime.
Difficutly implementing sliding pieces. Not sure on what approach to choose. (e.g using a for loop to loop through every direction which wouldnt be very efficient or opting for a more efficient approach such as magic bitboards or hyperbola quintessence.)