Solved and optimized piece bitboard access during makeMove()
Before, I had an if statements nested in a match case which was quite inefficient.

However now I have created a dictionary which has all the bitboards in it. Now whenever I access a bitboard it is done through the dictionary, which  optimizes the code alot due to its fast lookup speeds. I can also integrate this easily with my redundant mailbox representation, as I can call the pieces in each square the same as the corresponding key in the dictionary, which will also allow for fast access to bitboards of pieces that are being captured in the makeMove().