General Aesthetics + Log in page
Added Login Screen allowing the user to enter username and password.

TODO. 1) Connect to a database which can be used to verify the users username and password. 2) Create a new page for users to register. 3) Continue on board move generation. Currently 22 are generated at the start, when there should be 20. Also move representation is starting to become a larger issue. WIll need looking into.